[Hadoop] Move CredPropsUtil and dependencies from spark to hadoop module
Move CredPropsUtil, credential providers (AwsVendedTokenProvider,
GcsVendedTokenProvider, AbfsVendedTokenProvider), GenericCredentialProvider,
GenericCredential, CredScopedFileSystem, CredScopedFs, CredScopedKey,
UCHadoopConf, and ApiClientFactory from connectors/spark to connectors/hadoop
to define them as general APIs without Spark-specific dependencies.

Replace Spark-shaded Guava (ImmutableMap, Preconditions, Cache) with
standard Java (HashMap, Collections.unmodifiableMap, Objects.requireNonNull,
ConcurrentHashMap). Retain spark.UCHadoopConf as a thin subclass for
backward compatibility. Update build.sbt so spark depends on hadoop.
