+++ connectors/hadoop/src/main/java/io/unitycatalog/hadoop/auth/storage/AbfsVendedTokenProvider.java
@@ -1,10 +1,10 @@
-package io.unitycatalog.spark.auth.storage;
+package io.unitycatalog.hadoop.auth.storage;
 
 import io.unitycatalog.client.model.AzureUserDelegationSAS;
-import io.unitycatalog.spark.UCHadoopConf;
+import io.unitycatalog.hadoop.UCHadoopConf;
+import java.util.Objects;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.azurebfs.extensions.SASTokenProvider;
-import org.apache.hadoop.shaded.com.google.common.base.Preconditions;
 
 public class AbfsVendedTokenProvider extends GenericCredentialProvider implements SASTokenProvider {
   public static final String ACCESS_TOKEN_KEY = "fs.azure.sas.fixed.token";
@@ -22,17 +22,20 @@ public class AbfsVendedTokenProvider extends GenericCredentialProvider implement
         && conf.get(UCHadoopConf.AZURE_INIT_SAS_TOKEN_EXPIRED_TIME) != null) {
 
       String sasToken = conf.get(UCHadoopConf.AZURE_INIT_SAS_TOKEN);
-      Preconditions.checkNotNull(
+      Objects.requireNonNull(
           sasToken,
-          "Azure SAS token not set, please check " + "'%s' in hadoop configuration",
-          UCHadoopConf.AZURE_INIT_SAS_TOKEN);
+          String.format(
+              "Azure SAS token not set, please check '%s' in hadoop configuration",
+              UCHadoopConf.AZURE_INIT_SAS_TOKEN));
 
       long expiredTimeMillis = conf.getLong(UCHadoopConf.AZURE_INIT_SAS_TOKEN_EXPIRED_TIME, 0L);
-      Preconditions.checkState(
-          expiredTimeMillis > 0,
-          "Azure SAS token expired time must be greater than 0, please check '%s' in hadoop "
-              + "configuration",
-          UCHadoopConf.AZURE_INIT_SAS_TOKEN_EXPIRED_TIME);
+      if (expiredTimeMillis <= 0) {
+        throw new IllegalStateException(
+            String.format(
+                "Azure SAS token expired time must be greater than 0, please check '%s' in hadoop "
+                    + "configuration",
+                UCHadoopConf.AZURE_INIT_SAS_TOKEN_EXPIRED_TIME));
+      }
 
       return GenericCredential.forAzure(sasToken, expiredTimeMillis);
     } else {
@@ -45,7 +48,7 @@ public class AbfsVendedTokenProvider extends GenericCredentialProvider implement
     GenericCredential generic = accessCredentials();
 
     AzureUserDelegationSAS azureSAS = generic.temporaryCredentials().getAzureUserDelegationSas();
-    Preconditions.checkNotNull(azureSAS, "Azure SAS of generic credential cannot be null");
+    Objects.requireNonNull(azureSAS, "Azure SAS of generic credential cannot be null");
 
     return azureSAS.getSasToken();
   }
