+++ connectors/hadoop/src/main/java/io/unitycatalog/hadoop/auth/storage/AwsVendedTokenProvider.java
@@ -1,8 +1,8 @@
-package io.unitycatalog.spark.auth.storage;
+package io.unitycatalog.hadoop.auth.storage;
 
-import io.unitycatalog.spark.UCHadoopConf;
+import io.unitycatalog.hadoop.UCHadoopConf;
+import java.util.Objects;
 import org.apache.hadoop.conf.Configuration;
-import org.sparkproject.guava.base.Preconditions;
 import software.amazon.awssdk.auth.credentials.AwsCredentials;
 import software.amazon.awssdk.auth.credentials.AwsCredentialsProvider;
 import software.amazon.awssdk.auth.credentials.AwsSessionCredentials;
@@ -29,11 +29,12 @@ public class AwsVendedTokenProvider extends GenericCredentialProvider
 
       long expiredTimeMillis =
           conf.getLong(UCHadoopConf.S3A_INIT_CRED_EXPIRED_TIME, Long.MAX_VALUE);
-      Preconditions.checkState(
-          expiredTimeMillis > 0,
-          "Expired time %s must be greater than 0, " + "please check configure key '%s'",
-          expiredTimeMillis,
-          UCHadoopConf.S3A_INIT_CRED_EXPIRED_TIME);
+      if (expiredTimeMillis <= 0) {
+        throw new IllegalStateException(
+            String.format(
+                "Expired time %d must be greater than 0, please check configure key '%s'",
+                expiredTimeMillis, UCHadoopConf.S3A_INIT_CRED_EXPIRED_TIME));
+      }
 
       return GenericCredential.forAws(accessKey, secretKey, sessionToken, expiredTimeMillis);
     } else {
@@ -48,7 +49,7 @@ public class AwsVendedTokenProvider extends GenericCredentialProvider
     // Wrap the GenericCredential as an AwsCredentials.
     io.unitycatalog.client.model.AwsCredentials awsTempCred =
         generic.temporaryCredentials().getAwsTempCredentials();
-    Preconditions.checkNotNull(
+    Objects.requireNonNull(
         awsTempCred, "AWS temp credential of generic credentials cannot be null");
 
     return AwsSessionCredentials.builder()
