+++ connectors/hadoop/src/main/java/io/unitycatalog/hadoop/auth/storage/GcsVendedTokenProvider.java
@@ -1,12 +1,12 @@
-package io.unitycatalog.spark.auth.storage;
+package io.unitycatalog.hadoop.auth.storage;
 
 import com.google.cloud.hadoop.util.AccessTokenProvider;
 import io.unitycatalog.client.model.GcpOauthToken;
-import io.unitycatalog.spark.UCHadoopConf;
+import io.unitycatalog.hadoop.UCHadoopConf;
 import java.io.IOException;
 import java.time.Instant;
+import java.util.Objects;
 import org.apache.hadoop.conf.Configuration;
-import org.sparkproject.guava.base.Preconditions;
 
 public class GcsVendedTokenProvider extends GenericCredentialProvider
     implements AccessTokenProvider {
@@ -21,18 +21,20 @@ public class GcsVendedTokenProvider extends GenericCredentialProvider
   public GenericCredential initGenericCredential(Configuration conf) {
     if (conf.get(UCHadoopConf.GCS_INIT_OAUTH_TOKEN) != null) {
       String oauthToken = conf.get(UCHadoopConf.GCS_INIT_OAUTH_TOKEN);
-      Preconditions.checkNotNull(
+      Objects.requireNonNull(
           oauthToken,
-          "GCS OAuth token not set, please check '%s' in hadoop " + "configuration",
-          UCHadoopConf.GCS_INIT_OAUTH_TOKEN);
+          String.format(
+              "GCS OAuth token not set, please check '%s' in hadoop configuration",
+              UCHadoopConf.GCS_INIT_OAUTH_TOKEN));
 
       long expiredTimeMillis =
           conf.getLong(UCHadoopConf.GCS_INIT_OAUTH_TOKEN_EXPIRATION_TIME, Long.MAX_VALUE);
-      Preconditions.checkState(
-          expiredTimeMillis > 0,
-          "Expired time %s must be greater than 0, " + "please check configure key '%s'",
-          expiredTimeMillis,
-          UCHadoopConf.GCS_INIT_OAUTH_TOKEN_EXPIRATION_TIME);
+      if (expiredTimeMillis <= 0) {
+        throw new IllegalStateException(
+            String.format(
+                "Expired time %d must be greater than 0, please check configure key '%s'",
+                expiredTimeMillis, UCHadoopConf.GCS_INIT_OAUTH_TOKEN_EXPIRATION_TIME));
+      }
 
       return GenericCredential.forGcs(oauthToken, expiredTimeMillis);
     } else {
@@ -45,10 +47,10 @@ public class GcsVendedTokenProvider extends GenericCredentialProvider
     GenericCredential generic = accessCredentials();
 
     GcpOauthToken gcpToken = generic.temporaryCredentials().getGcpOauthToken();
-    Preconditions.checkNotNull(gcpToken, "GCS OAuth token of generic credential cannot be null");
+    Objects.requireNonNull(gcpToken, "GCS OAuth token of generic credential cannot be null");
 
     String tokenValue = gcpToken.getOauthToken();
-    Preconditions.checkNotNull(tokenValue, "GCS OAuth token value cannot be null");
+    Objects.requireNonNull(tokenValue, "GCS OAuth token value cannot be null");
 
     Long expirationMillis = generic.temporaryCredentials().getExpirationTime();
     Instant expirationInstant =
