+++ connectors/hadoop/src/main/java/io/unitycatalog/hadoop/auth/storage/GenericCredential.java
@@ -1,4 +1,4 @@
-package io.unitycatalog.spark.auth.storage;
+package io.unitycatalog.hadoop.auth.storage;
 
 import io.unitycatalog.client.internal.Clock;
 import io.unitycatalog.client.model.AwsCredentials;
