+++ connectors/hadoop/src/test/java/io/unitycatalog/hadoop/auth/CredPropsUtilTest.java
@@ -1,4 +1,4 @@
-package io.unitycatalog.spark.auth;
+package io.unitycatalog.hadoop.auth;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -14,7 +14,7 @@ import org.junit.jupiter.api.Test;
 /**
  * Verifies that {@link CredPropsUtil} saves the original {@code fs.<scheme>.impl} values under
  * {@code fs.<scheme>.impl.original} before overriding them with {@link
- * io.unitycatalog.spark.fs.CredScopedFileSystem}, so that the real delegate can be restored in
+ * io.unitycatalog.hadoop.fs.CredScopedFileSystem}, so that the real delegate can be restored in
  * {@code CredScopedFileSystem#newFileSystem}.
  */
 class CredPropsUtilTest {
