+++ connectors/spark/src/test/java/io/unitycatalog/spark/GCSCredentialTestFileSystem.java
@@ -2,7 +2,7 @@ package io.unitycatalog.spark;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import io.unitycatalog.spark.auth.storage.GcsVendedTokenProvider;
+import io.unitycatalog.hadoop.auth.storage.GcsVendedTokenProvider;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.Path;
