+++ connectors/spark/src/test/java/io/unitycatalog/spark/S3CredentialTestFileSystem.java
@@ -2,7 +2,7 @@ package io.unitycatalog.spark;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import io.unitycatalog.spark.auth.storage.AwsVendedTokenProvider;
+import io.unitycatalog.hadoop.auth.storage.AwsVendedTokenProvider;
 import java.io.IOException;
 import java.net.URI;
 import java.util.Objects;
