+++ connectors/spark/src/test/java/io/unitycatalog/spark/auth/storage/AbfsCredRenewITTest.java
@@ -2,6 +2,7 @@ package io.unitycatalog.spark.auth.storage;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
+import io.unitycatalog.hadoop.auth.storage.AbfsVendedTokenProvider;
 import io.unitycatalog.server.service.credential.azure.ADLSStorageConfig;
 import io.unitycatalog.server.service.credential.azure.AzureCredential;
 import io.unitycatalog.server.service.credential.azure.AzureCredentialGenerator;
