+++ connectors/spark/src/test/java/io/unitycatalog/spark/auth/storage/AwsCredScopedFsRenewITTest.java
@@ -5,7 +5,7 @@ import java.util.Map;
 
 /**
  * Re-runs the AWS credential-renewal test with {@code credScopedFs.enabled=true} to verify that
- * {@link io.unitycatalog.spark.fs.CredScopedFileSystem} wraps the delegate transparently and
+ * {@link io.unitycatalog.hadoop.fs.CredScopedFileSystem} wraps the delegate transparently and
  * credential renewal still works correctly end-to-end.
  *
  * <p>Sets {@code fs.s3.impl} and {@code fs.s3a.impl} at the session level to {@link
