+++ connectors/spark/src/test/java/io/unitycatalog/spark/auth/storage/AwsVendedTokenProviderTest.java
@@ -6,6 +6,7 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import io.unitycatalog.client.api.TemporaryCredentialsApi;
 import io.unitycatalog.client.model.AwsCredentials;
 import io.unitycatalog.client.model.TemporaryCredentials;
+import io.unitycatalog.hadoop.auth.storage.AwsVendedTokenProvider;
 import io.unitycatalog.spark.UCHadoopConf;
 import java.io.IOException;
 import java.net.URI;
