+++ connectors/spark/src/test/java/io/unitycatalog/spark/auth/storage/BaseCredRenewITTest.java
@@ -7,6 +7,7 @@ import io.delta.tables.DeltaTable;
 import io.unitycatalog.client.internal.Clock;
 import io.unitycatalog.client.model.CreateCatalog;
 import io.unitycatalog.client.model.CreateSchema;
+import io.unitycatalog.hadoop.fs.CredScopedFileSystem;
 import io.unitycatalog.server.base.BaseCRUDTest;
 import io.unitycatalog.server.base.ServerConfig;
 import io.unitycatalog.server.base.catalog.CatalogOperations;
@@ -16,7 +17,6 @@ import io.unitycatalog.server.service.credential.CredentialContext;
 import io.unitycatalog.spark.CredentialTestFileSystem;
 import io.unitycatalog.spark.UCHadoopConf;
 import io.unitycatalog.spark.UCSingleCatalog;
-import io.unitycatalog.spark.fs.CredScopedFileSystem;
 import java.io.File;
 import java.net.URI;
 import java.time.Duration;
