+++ connectors/spark/src/test/java/io/unitycatalog/spark/auth/storage/BaseTokenProviderTest.java
@@ -10,6 +10,8 @@ import io.unitycatalog.client.internal.Clock;
 import io.unitycatalog.client.model.PathOperation;
 import io.unitycatalog.client.model.TableOperation;
 import io.unitycatalog.client.model.TemporaryCredentials;
+import io.unitycatalog.hadoop.auth.storage.GenericCredential;
+import io.unitycatalog.hadoop.auth.storage.GenericCredentialProvider;
 import io.unitycatalog.spark.UCHadoopConf;
 import java.time.Duration;
 import java.util.UUID;
@@ -38,7 +40,7 @@ public abstract class BaseTokenProviderTest<T extends GenericCredentialProvider>
   public void before() {
     clockName = UUID.randomUUID().toString();
     clock = Clock.getManualClock(clockName);
-    GenericCredentialProvider.globalCache.invalidateAll();
+    GenericCredentialProvider.globalCache.clear();
   }
 
   @AfterEach
@@ -46,7 +48,7 @@ public abstract class BaseTokenProviderTest<T extends GenericCredentialProvider>
     Clock.removeManualClock(clockName);
     clock = null;
     clockName = null;
-    GenericCredentialProvider.globalCache.invalidateAll();
+    GenericCredentialProvider.globalCache.clear();
   }
 
   @Test
@@ -308,7 +310,7 @@ public abstract class BaseTokenProviderTest<T extends GenericCredentialProvider>
     assertThat(expectedSize).isEqualTo(creds.length);
     assertThat(GenericCredentialProvider.globalCache.size()).isEqualTo(expectedSize);
     for (TemporaryCredentials cred : creds) {
-      assertThat(GenericCredentialProvider.globalCache.asMap().values())
+      assertThat(GenericCredentialProvider.globalCache.values())
           .contains(new GenericCredential(cred));
     }
   }
