+++ connectors/spark/src/test/java/io/unitycatalog/spark/fs/CredScopedFileSystemCacheTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 
+import io.unitycatalog.hadoop.fs.CredScopedFileSystem;
+import io.unitycatalog.hadoop.fs.CredScopedKey;
 import io.unitycatalog.spark.UCHadoopConf;
 import java.net.URI;
 import org.apache.hadoop.conf.Configuration;
